Add Fluid class and FluidPhysics for fluid property calculations
